+++ tests/integration/constant.cpp
@@ -1,11 +1,11 @@
 #include <iostream>
 
 extern "C" {
-double get_constant(double);
+float get_constant(float);
 }
 
 int main() {
-  double value = get_constant(3.0);
+  float value = get_constant(3.0);
   printf("get_constant(3.0): %f\n", value);
 
   return 0;
